+++ .gitignore
@@ -50,3 +50,4 @@ lib-cov
 
 # Stores VSCode versions used for testing VSCode extensions
 .vscode-test
+.idea
