+++ app/calendars/types.tsx
@@ -5,5 +5,14 @@ export interface Meeting {
   timezone: number
   startDate: string
   endDate: string
+  schedule: {
+    monday: any[]
+    tuesday: any[]
+    wendsday: any[]
+    thursday: any[]
+    friday: any[]
+    saturday: any[]
+    sunday: any[]
+  }
   timeslots: string[]
 }
